Move XML stuff into lib
