+++ src/lib.js
@@ -1,4 +1,4 @@
-let aws, ini
+let aws, ini, xml
 
 // AWS-flavored JSON stuff
 function marshaller (method, obj, awsjsonSetting) {
@@ -80,4 +80,54 @@ function useAWS () {
   return true
 }
 
-module.exports = { awsjson, exists, loadAwsConfig, readConfig, tidyQuery, useAWS }
+// XML stuff
+let textNodeName = '#text'
+/* istanbul ignore next */
+function instantiateXml () {
+  if (xml) return
+  // eslint-disable-next-line
+  let vendor = require('./_vendor/xml')
+  // The following was pulled directly from AWS's implementations of `fast-xml-parser` in SDKv3
+  xml = {
+    parser: new vendor.XMLParser({
+      attributeNamePrefix: '',
+      htmlEntities: true,
+      ignoreAttributes: false,
+      ignoreDeclaration: true,
+      parseTagValue: false,
+      trimValues: false,
+      tagValueProcessor: (_, val) => (val.trim() === '' && val.includes('\n') ? '' : undefined),
+    }),
+    builder: new vendor.XMLBuilder(),
+  }
+  xml.parser.addEntity('#xD', '\r')
+  xml.parser.addEntity('#10', '\n')
+  xml.parser.getValueFromTextNode = vendor.getValueFromTextNode
+}
+function buildXml (obj) {
+  instantiateXml()
+  return xml.builder.build(obj)
+}
+function parseXml (body) {
+  instantiateXml()
+  let parsed = xml.parser.parse(body)
+  let key = Object.keys(parsed)[0]
+  let payloadToReturn = parsed[key]
+  /* istanbul ignore next */ // TODO remove + test
+  if (payloadToReturn[textNodeName]) {
+    payloadToReturn[key] = payloadToReturn[textNodeName]
+    delete payloadToReturn[textNodeName]
+  }
+  return xml.parser.getValueFromTextNode(payloadToReturn)
+}
+
+module.exports = {
+  awsjson,
+  exists,
+  loadAwsConfig,
+  readConfig,
+  tidyQuery,
+  useAWS,
+  buildXml,
+  parseXml,
+}
